WIP: Test replace one increment batch with add
This is just a test to see if we can replace the increment batch
(4 instructions that make up the increment) with a
single instruction (add in this case).
Currently we only replace the final add instruction with our add.
We need to find a way to remove previous 3 instructions without
breaking the next increments.