WIP: Count allocas with another pass
Increments have the load-add-store pattern after the initial
sequence of allocas and stores.
It seems that the number of allocas and stores is equal to the
number of function arguments + local variables.
We could grab the number of function arguments from our main pass,
but it seems that LLVM doesn't have a function that grabs the
number of locally declared variables.
Thats why we have another pass that just checks that.